backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
accelerate==1.12.0
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==25.1.0
Authlib==1.6.5
boto3==1.40.49
botocore==1.40.49
botocore-stubs==1.40.33
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==46.0.3
deprecation==2.1.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.118.0
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.0
filelock==3.20.0
flatbuffers==25.9.23
fsspec==2025.10.0
greenlet==3.2.4
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
magika==0.6.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
networkx==3.6
numpy==2.3.3
onnxruntime==1.23.1
packaging==25.0
pillow==12.0.0
portalocker==3.2.0
protobuf==6.32.1
psutil==7.2.1
psycopg2==2.9.10
pwdlib==0.2.1
pycparser==2.23
pydantic==2.11.10
pydantic-settings==2.11.0
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
PyMuPDF==1.27.2.2
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.1
regex==2025.11.3
requests==2.32.5
rich==14.1.0
rich-toolkit==0.15.1
rignore==0.7.0
s3transfer==0.14.0
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
sentence-transformers==5.1.2
sentry-sdk==2.39.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.1+cu128
torchvision==0.24.1+cu128
tqdm==4.67.1
transformers==4.57.3
typer==0.19.2
types-awscrt==0.27.6
types-boto3==1.40.49
types-boto3-full==1.40.48
types-s3transfer==0.13.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
validators==0.35.0
watchfiles==1.1.0
websockets==15.0.1
